+++ ChangeLog.md
@@ -29,6 +29,9 @@
         .build())
     ```
 
+* **[breaking]** Replaces `Vector:create<#x>()` with `Vector<#x>.new()` (in
+  `lib/container`) for consistency with other `@value` construction.
+
 * **[new]** Adds `LessThan2` `@value interface` to `lib/util`.
 
 * **[new]** Implements `Sorting` functions in `lib/container` that use  the new